fix: moved util file to common file and added todo
