+++ packages/nextjs/utils/scaffold-stark/common.ts
@@ -26,3 +26,14 @@ export function isJsonString(str: string) {
     return false;
   }
 }
+
+export function isValidContractArgs(
+  args: unknown,
+  expectedLength: number
+): boolean {
+  return (
+    Array.isArray(args) &&
+    args.length === expectedLength &&
+    args.every((arg) => arg !== undefined && arg !== null && arg !== "")
+  );
+}
